requirements.txt
Django==1.8
django-extensions==1.6.1
django-filter==0.12.0
djangorestframework==3.3.2
Markdown==2.6.5
pydot==1.0.2
pyparsing==1.5.7
six==1.10.0
